pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "blobmodel"
version = "1.0.0"
description = "Two dimensional model of propagating blobs"
authors = [
  { name = "gregordecristoforo", email = "gregor.decristoforo@gmail.com" }
]
license = { file = "LICENSE" }
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy",
    "xarray",
    "tqdm",
    "nptyping",
    "matplotlib",
    "netcdf4",
    "pytest-cov",
    "black",
    "mypy",
    "sphinx",
    "sphinx-rtd-theme",
]

[tool.setuptools]
packages = ["blobmodel"]
